[quant][refactor] Tests for torch.jit.quantized
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
